This extends the usability of the Kolmogorov-Smirnov test
from the Classic DoublesSketch to KllDoublesSketch and the
KllFloatsSketch.

In order to do this I needed to add "getNormalizedRankError(boolean) to
the root interface QuantilesAPI. All of the sketches (except REQ) had
this anyway. In REQ I just throw an unsupported operation exception,
because it makes no sense for REQ.

This also means the KolmogorovSmirnov class and its test move to the
"quantilescommon" package.
