+++ src/main/java/org/apache/datasketches/quantiles/DoublesSketch.java
@@ -266,15 +266,7 @@ public abstract class DoublesSketch implements QuantilesDoublesAPI {
   @Override
   public abstract long getN();
 
-  /**
-   * Gets the approximate rank error of this sketch normalized as a fraction between zero and one.
-   * The epsilon returned is a best fit to 99 percent confidence empirically measured max error
-   * in thousands of trials.
-   * @param pmf if true, returns the "double-sided" normalized rank error for the getPMF() function.
-   * Otherwise, it is the "single-sided" normalized rank error for all the other queries.
-   * @return if pmf is true, returns the normalized rank error for the getPMF() function.
-   * Otherwise, it is the "single-sided" normalized rank error for all the other queries.
-   */
+  @Override
   public double getNormalizedRankError(final boolean pmf) {
     return getNormalizedRankError(k_, pmf);
   }
