+++ src/main/java/org/apache/datasketches/quantilescommon/QuantilesAPI.java
@@ -226,6 +226,17 @@ public interface QuantilesAPI {
   long getN();
 
   /**
+   * Gets the approximate rank error of this sketch normalized as a fraction between zero and one.
+   * The epsilon returned is a best fit to 99 percent confidence empirically measured max error
+   * in thousands of trials.
+   * @param pmf if true, returns the "double-sided" normalized rank error for the getPMF() function.
+   * Otherwise, it is the "single-sided" normalized rank error for all the other queries.
+   * @return if pmf is true, returns the "double-sided" normalized rank error for the getPMF() function.
+   * Otherwise, it is the "single-sided" normalized rank error for all the other queries.
+   */
+  double getNormalizedRankError(boolean pmf);
+
+  /**
    * Gets the number of quantiles retained by the sketch.
    * @return the number of quantiles retained by the sketch
    */
